refactor: Player의 HP와 MP를 검증하는 로직을 각각의 객체의 위임함
