+++ src/main/java/bossmonster/domain/player/PlayerHpAndMp.java
@@ -3,43 +3,37 @@ package bossmonster.domain.player;
 import static bossmonster.domain.player.constant.PlayerOption.*;
 
 public class PlayerHpAndMp {
-    private final int hp;
-    private final int mp;
+    private final PlayerHp hp;
+    private final PlayerMp mp;
 
     public PlayerHpAndMp(
             int hp,
             int mp
     ) {
-        validateHp(hp);
-        validateMp(mp);
         validateHpAndMpSum(hp + mp);
-        this.hp = hp;
-        this.mp = mp;
+        this.hp = convertToPlayerHp(hp);
+        this.mp = convertToPlayerMp(mp);
     }
 
-    public int getHp() {
+    public PlayerHp getHp() {
         return hp;
     }
 
-    public int getMp() {
+    public PlayerMp getMp() {
         return mp;
-    }
-
-    private void validateHp(int hp) {
-        if (PLAYER_MINIMUM_HP_LIMIT > hp || PLAYER_MAXIMUM_HP_LIMIT < hp) {
-            throw new IllegalArgumentException("유효하지 않은 플레이어 HP입니다.");
-        }
-    }
-
-    private void validateMp(int mp) {
-        if (PLAYER_MINIMUM_MP_LIMIT > mp || PLAYER_MAXIMUM_MP_LIMIT < mp) {
-            throw new IllegalArgumentException("유효하지 않은 플레이어 MP입니다.");
-        }
     }
 
     private void validateHpAndMpSum(int hpAndMpSum) {
         if (hpAndMpSum != PLAYER_HP_AND_MP_SUM) {
             throw new IllegalArgumentException("플레이어의 HP와 MP의 합이 유효하지 않습니다.");
         }
+    }
+
+    private PlayerHp convertToPlayerHp(int hp) {
+        return new PlayerHp(hp);
+    }
+
+    private PlayerMp convertToPlayerMp(int mp) {
+        return new PlayerMp(mp);
     }
 }
