+++ src/main/java/bossmonster/domain/player/PlayerMp.java
@@ -1,10 +1,14 @@
 package bossmonster.domain.player;
 
+import static bossmonster.domain.player.constant.PlayerOption.PLAYER_MAXIMUM_MP_LIMIT;
+import static bossmonster.domain.player.constant.PlayerOption.PLAYER_MINIMUM_MP_LIMIT;
+
 public class PlayerMp {
     private final int maximumMp;
     private int currentMp;
 
     public PlayerMp(int mp) {
+        validateMp(mp);
         this.maximumMp = mp;
         this.currentMp = mp;
     }
@@ -18,10 +22,7 @@ public class PlayerMp {
     }
 
     public boolean hasEnoughMpForAttack(int enoughMp) {
-        if (enoughMp <= currentMp) {
-            return true;
-        }
-        return false;
+        return enoughMp <= currentMp;
     }
 
     public void increaseCurrentMp(int increaseMp) {
@@ -33,5 +34,11 @@ public class PlayerMp {
 
     public void decreaseCurrentMp(int decreaseMp) {
         currentMp -= decreaseMp;
+    }
+
+    private void validateMp(int mp) {
+        if (PLAYER_MINIMUM_MP_LIMIT > mp || PLAYER_MAXIMUM_MP_LIMIT < mp) {
+            throw new IllegalArgumentException("유효하지 않은 플레이어 MP입니다.");
+        }
     }
 }
